+++ Sources/VComponents/Components/Inputs/TextView/VTextViewUIModel.swift
@@ -62,7 +62,7 @@ public struct VTextViewUIModel {
     public var borderColors: StateColors = .clearColors
 
     // MARK: Properties - Header
-    /// Header title text line type. Set to `singleLine`.
+    /// Header title text line type. Set to `multiline` with `leading` alignment and `1...2` lines.
     public var headerTitleTextLineType: TextLineType = GlobalUIModel.Common.headerTitleTextLineType
 
     /// Header title text colors.