+++ Sources/VComponents/Components/Item Pickers/Wheel Picker/VWheelPickerUIModel.swift
@@ -27,7 +27,7 @@ public struct VWheelPickerUIModel {
     public var backgroundColors: StateColors = .init(ColorBook.layer)
 
     // MARK: Properties - Header
-    /// Header title text line type. Set to `singleLine`.
+    /// Header title text line type. Set to `multiline` with `leading` alignment and `1...2` lines.
     public var headerTitleTextLineType: TextLineType = GlobalUIModel.Common.headerTitleTextLineType
 
     /// Header title text colors.