+++ Sources/VComponents/Components/Modals/Alert/VAlertUIModel.swift
@@ -67,8 +67,14 @@ public struct VAlertUIModel {
     }
 
     // MARK: Properties - Title
-    /// Title text line type. Set to `singleLine`.
-    public var titleTextLineType: TextLineType = .singleLine
+    /// Title text line type. Set to `multiline` with `center` alignment and `1...2` lines.
+    public var titleTextLineType: TextLineType = {
+        if #available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *) {
+            return .multiLine(alignment: .center, lineLimit: 1...2)
+        } else {
+            return .multiLine(alignment: .center, lineLimit: 2)
+        }
+    }()
 
     /// Title text color.
     public var titleTextColor: Color = ColorBook.primary